+++ backend/smody/src/main/java/com/woowacourse/smody/push/controller/PushNotificationController.java
@@ -5,7 +5,7 @@ import com.woowacourse.smody.auth.login.LoginMember;
 import com.woowacourse.smody.auth.login.RequiredLogin;
 import com.woowacourse.smody.push.dto.MentionNotificationRequest;
 import com.woowacourse.smody.push.dto.PushNotificationResponse;
-import com.woowacourse.smody.push.service.PushNotificationService;
+import com.woowacourse.smody.push.service.PushNotificationApiService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -22,19 +22,20 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class PushNotificationController {
 
-    private final PushNotificationService pushNotificationService;
+    private final PushNotificationApiService pushNotificationApiService;
 
+    //TODO: api 수정(/push-notifications/me)
     @GetMapping
     @RequiredLogin
-    public ResponseEntity<List<PushNotificationResponse>> searchNotificationsOfMine(
+    public ResponseEntity<List<PushNotificationResponse>> searchNotificationsByMe(
             @LoginMember TokenPayload tokenPayload) {
-        return ResponseEntity.ok(pushNotificationService.searchNotificationsOfMine(tokenPayload));
+        return ResponseEntity.ok(pushNotificationApiService.searchNotificationsByMe(tokenPayload));
     }
 
     @DeleteMapping("{id}")
     @RequiredLogin
     public ResponseEntity<Void> deleteNotification(@PathVariable Long id) {
-        pushNotificationService.delete(id);
+        pushNotificationApiService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
 
@@ -42,7 +43,7 @@ public class PushNotificationController {
     @RequiredLogin
     public ResponseEntity<Void> saveNotification(@LoginMember TokenPayload tokenPayload,
                                                  @RequestBody MentionNotificationRequest mentionNotificationRequest) {
-        pushNotificationService.saveNotification(tokenPayload, mentionNotificationRequest);
+        pushNotificationApiService.saveNotification(tokenPayload, mentionNotificationRequest);
         return ResponseEntity.ok().build();
     }
 }
